Added existance check to the delete search functions

+++ controllers/searchEngineController.js
@@ -3,10 +3,14 @@ const { pool } = require("../config/db");
 
 // Returning past searched values
 const getPastSearches = async (req, res) => {
+    const { userId } = req.user;
+    if (!userId) {
+        return res.status(400).json({ success: false, error: "userId is required" });
+    };
     try {
-        const query = "SELECT * FROM searches";
+        const query = "SELECT * FROM searches WHERE user_id = $1 ORDER BY created_at DESC LIMIT 10";
 
-        const result = await pool.query(query);
+        const result = await pool.query(query, [userId]);
 
         if (result.rows.length === 0) {
             return res.status(404).json({ success: false, error: "No past searches found" });
@@ -27,16 +31,51 @@ const deleteSearch = async (req, res) => {
         if (!id) {
             return res.status(400).json({ success: false, error: "Search term Id is required"})
         }
+
+        // ckeck if the search query exists
+        const checkQuery = "SELECT * FROM searches WHERE search_id = $1";
+        const checkResult = await pool.query(checkQuery, [id]);
+
+        if (checkResult.rows.length > 0) {
+            return res.status(404).json({ success: false, error: "Search query not found" });
+        }
+
         const query = "DELETE FROM searches WHERE search_id = $1";
 
         const result = await pool.query(query, [id]);
 
         return res.status(200).json({ success: true, message: "Search deleted successfully" });
-
     }
     catch (err) {
         console.log(err);
         return res.status(500).json({ success: false, error: "An error occurred with the delete search query" });
+    }
+}
+
+// Delete all past searches using userId
+const deleteAllSearches = async (req, res) => {
+    const {userId} = req.user;
+
+    if (!userId) {
+        return res.status(400).json({ success: false, error: "userId is required" });
+    };
+
+    try {
+        // Check if the user has any searches
+        const checkQuery = "SELECT * FROM searches WHERE user_id = $1";
+        const checkResult = await pool.query(checkQuery, [userId]);
+
+        if (checkResult.rows.length === 0) {
+            return res.status(404).json({ success: false, error: "No searches found" });
+        }
+        
+        const query = "DELETE FROM searches WHERE user_id = $1";
+
+        const result = await pool.query(query, [userId]);
+        return res.status(200).json({ success: true, message: "All searches deleted successfully" });
+    } catch (err) {
+        console.log(err);
+        return res.status(500).json({ success: false, error: "An error occurred with the delete all searches query" });
     }
 }
 
@@ -94,4 +133,4 @@ const searchEngine = async (req, res) => {
 }
 
 
-module.exports = { searchEngine, deleteSearch, getPastSearches }
+module.exports = { searchEngine, deleteSearch, getPastSearches, deleteAllSearches }
